Client and ClientContact routes/services. Job and Address seeds. etc

+++ server/user/userRoutes.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash')
 const service = require('./userService');
-const { handleInitialSuccess, handleInitialFailure } = require('../utils/routeHelpers');
+const { handleInitialSuccess, handleInitialFailure, permissionError, checkPermission } = require('../utils/routeHelpers');
 
 const routes = [];
 
@@ -12,13 +12,19 @@ routes.push(
     path: '/api/v1/users',
     async handler(req, h) {
       const { query } = req;
-      try {
-        const data = !query ? 
-            await service.getUsers() : 
+      const { role } = req.auth.credentials;
+      const allowedRoles = ['Admin', 'Manager', 'Customer Service'];
+      if(checkPermission(req, allowedRoles)) {
+        try {
+          const data = !query ?
+            await service.getUsers() :
             await service.getUser(query);
-        return handleInitialSuccess(h, data);
-      } catch(error) {
-        return handleInitialFailure(h, 'Failed to retrieve user(s)');
+          return handleInitialSuccess(h, data);
+        } catch (error) {
+          return handleInitialFailure(h, 'Failed to retrieve user(s)');
+        }
+      } else {
+        permissionError(h, role);
       }
     },
     config: {
@@ -30,11 +36,17 @@ routes.push(
     path: '/api/v1/users/{id}',
     async handler(req, h) {
       const { id } = req.params;
-      try {
-        const data = await service.getUserById(id);
-        return handleInitialSuccess(h, data);
-      } catch (error) {
-        return handleInitialFailure(h, `Failed to retrieve user with id: ${id}`);
+      const { role } = req.auth.credentials;
+      const allowedRoles = ['Admin', 'Manager', 'Customer Service'];
+      if (checkPermission(req, allowedRoles)) {
+        try {
+          const data = await service.getUserById(id);
+          return handleInitialSuccess(h, data);
+        } catch (error) {
+          return handleInitialFailure(h, `Failed to retrieve user with id: ${id}`);
+        }
+      } else {
+        permissionError(h, role);
       }
     },
     config: {
@@ -46,12 +58,21 @@ routes.push(
     path: '/api/v1/users',
     async handler(req, h) {
       const data = req.payload;
-      try {
-        const data = await service.createUser(data);
-        return handleInitialSuccess(h, data);
-      } catch (error) {
-        return handleInitialFailure(h, `Failed to create user with email: ${data.email}`);
+      const { role } = req.auth.credentials;
+      const allowedRoles = ['Admin'];
+      if (checkPermission(req, allowedRoles)) {
+        try {
+          const data = await service.createUser(data);
+          return handleInitialSuccess(h, data);
+        } catch (error) {
+          return handleInitialFailure(h, `Failed to create user with email: ${data.email}`);
+        }
+      } else {
+        permissionError(h, role);
       }
+    },
+    config: {
+      tags: ['api', 'v1', 'users', 'create']
     }
   },
   {
@@ -60,12 +81,21 @@ routes.push(
     async handler(req, h) {
       const data = req.payload;
       const { id } = req.params;
-      try {
-        const data = await service.updateUser(id, data);
-        return handleInitialSuccess(h, data);
-      } catch (error) {
-        return handleInitialFailure(h, `Failed to update user with email: ${data.email}`);
+      const { role } = req.auth.credentials;
+      const allowedRoles = ['Admin'];
+      if (checkPermission(req, allowedRoles)) {
+        try {
+          const data = await service.updateUser(id, data);
+          return handleInitialSuccess(h, data);
+        } catch (error) {
+          return handleInitialFailure(h, `Failed to update user with email: ${data.email}`);
+        }
+      } else {
+        permissionError(h, role);
       }
+    },
+    config: {
+      tags: ['api', 'v1', 'users', 'update']
     }
   },
   {
@@ -73,15 +103,21 @@ routes.push(
     path: '/api/v1/users/{id}',
     async handler(req, h) {
       const { id } = req.params;
-      try {
-        const data = await service.deleteUser(id);
-        return handleInitialSuccess(h, data);
-      } catch (error) {
-        return handleInitialFailure(h, `Failed to delete user with id: ${id}`);
+      const { role } = req.auth.credentials;
+      const allowedRoles = ['Admin'];
+      if (checkPermission(req, allowedRoles)) {
+        try {
+          const data = await service.deleteUser(id);
+          return handleInitialSuccess(h, data);
+        } catch (error) {
+          return handleInitialFailure(h, `Failed to delete user with id: ${id}`);
+        }
+      } else {
+        permissionError(h, role);
       }
     },
     config: {
-      auth: false
+      tags: ['api', 'v1', 'users', 'delete']
     }
   }
 )
